Move profiler -> prototype
